emulationService's added

+++ src/main/java/model/Horse.java
@@ -7,5 +7,5 @@ public class Horse {
     private final String name;
     private final Rider rider;
     private final Breed breed;
-    private final int averageSpeed;
+    private final double averageSpeed;
 }
